// clients/js/src/generated/instructions/updateAssetV1.ts
export type UpdateAssetV1InstructionAccounts = {
  /** Candy Guard account. */
  candyGuard?: PublicKey | Pda;
  /**
   * Candy Machine program account.
   *
   */

  candyMachineProgram?: PublicKey | Pda;
  /** Candy machine account. */
  candyMachine: PublicKey | Pda;
  /**
   * Candy Machine authority account.
   *
   */

  candyMachineAuthorityPda?: PublicKey | Pda;
  /** Payer for the mint (SOL) fees. */
  payer?: Signer;
  /** Minter account for validation and non-SOL fees. */
  minter?: Signer;
  /** Optionally mint to different owner */
  owner?: PublicKey | Pda;
  /**
   * Mint account of the NFT. The account will be initialized if necessary.
   *
   * Must be a signer if:
   * * the nft_mint account does not exist.
   *
   */

  asset: PublicKey | Pda | Signer;
  /**
   * Mint account of the collection NFT.
   *
   */

  collection: PublicKey | Pda;
  /**
   * Token Metadata program.
   *
   * Token Metadata program.
   *
   */

  mplCoreProgram?: PublicKey | Pda;
  /**
   * SPL Token program.
   * System program.
   */

  systemProgram?: PublicKey | Pda;
  /**
   * Instructions sysvar account.
   *
   */

  sysvarInstructions?: PublicKey | Pda;
  /**
   * SlotHashes sysvar cluster data.
   *
   */

  recentSlothashes?: PublicKey | Pda;
};